requirements.txt
beautifulsoup4==4.13.4
chromadb==1.0.13
fastapi
httpx
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.67
langchain-mistralai==0.2.10
langchain-tavily==0.2.5
langgraph==0.5.0
mistralai==1.9.1
pandas
pypdf2==3.0.1
python-dotenv==1.1.1
requests==2.32.4
streamlit
tavily-python
uvicorn